Move FingerprintPatcher out of Internal namespace
This class is needed to write patchers for non-Item/Property entities
with fingerprints, so let’s make it public. Also, like other patchers in
the library it now accepts any Diff, not just an EntityDiff; this makes
it more useful to patch fingerprints not connected to proper entities,
and also reduces the dependency upon EntityDiff’s label / description /
aliases getters, which are already marked with a FIXME because not all
entities have labels, descriptions or aliases.
